Rename request -> api, make GitOpts a value
